refact: 프로모션 타이머를 features로 분리
